+++ src/Statement.ts
@@ -1,79 +1,17 @@
+import {
+  Balance,
+  EffectBlock,
+  ConditionBlock,
+  StatementInterface,
+  Context,
+  ConditionKey,
+  ConditionResolver
+} from './definitions'
+
 const reDelimiters = /\${([^}])*}/g;
 const trim = / +(?= )|^\s+|\s+$/g;
 
-type StatementEffect = 'allow' | 'deny';
-//type PrincipalAttributtes = 'principal' | 'notPrincipal';
-type StatementPattern = string;
-
-interface PrincipalMap {
-  [key: string]: string[] | string;
-}
-
-interface PrincipalBlock {
-  principal?: PrincipalMap | string; // '*'
-}
-
-interface NotPrincipal {
-  notPrincipal?: PrincipalMap | string; // '*'
-}
-
-interface Action {
-  action: string[] | string;
-}
-
-interface NotAction {
-  notAction: string[] | string;
-}
-
-interface Resource {
-  resource: string[] | string;
-}
-
-interface NotResource {
-  notResource: string[] | string;
-}
-
-interface Condition {
-  [key: string]: string | number |boolean;
-}
-
-/*interface MessageWithText extends MessageBasics {
-  text: string;
-}
-interface MessageWithAttachment extends MessageBasics {
-  attachment: Attachment;
-}
-type Message = MessageWithText | MessageWithAttachment;
-
-type Message = MessageWithText | MessageWithAttachment | (MessageWithText & MessageWithAttachment);*/
-
-interface StatementConditions {
-  [key: string]: Condition;
-}
-
-interface Balance {
-  start: number;
-  end: number;
-  pre: string;
-  body: string;
-  post: string;
-}
-
-type StatementConfig = {
-  sid?: StatementPattern;
-  effect?: StatementEffect;
-  resource: StatementPattern[] | StatementPattern;
-  action: StatementPattern[] | StatementPattern;
-  condition?: StatementConditions;
-};
-
-type Resolver = (data: any, expected: any) => boolean;
-
-interface ConditionResolver {
-  [key: string]: Resolver;
-}
-
-export function getValueFromPath(data: any, path: string): any {
+export function getValueFromPath(data: any, path: string): ConditionKey {
   const value = path.split('.').reduce((key, nextChild) => {
     return data[key] ? data[key][nextChild] : undefined;
   });
@@ -85,7 +23,7 @@ export function getValueFromPath(data: any, path: string): any {
 
 const specialTrim = (str: string): string => str.replace(trim, '');
 
-export function applyContext(str: string, context?: object): string {
+export function applyContext(str: string, context?: Context): string {
   if (!context) return str;
 
   return specialTrim(
@@ -247,46 +185,47 @@ export class Matcher {
   }
 }
 
-export class Statement {
-  effect: StatementEffect;
-  private resource: StatementPattern[];
-  private action: StatementPattern[];
-  private readonly condition?: StatementConditions;
+class Statement {
+  effect: EffectBlock;
+  protected readonly condition?: ConditionBlock;
 
   constructor({
     effect = 'allow',
-    resource,
-    action,
     condition,
-  }: StatementConfig) {
+  }: StatementInterface) {
     this.effect = effect;
-    this.resource = typeof resource === 'string' ? [resource] : resource;
-    this.action = typeof action === 'string' ? [action] : action;
     this.condition = condition;
   }
 
-  matches(
-    action: string,
-    resource: string,
-    context?: object,
+  matchConditions(
+    context?: Context,
     conditionResolvers?: ConditionResolver
   ): boolean {
     return (
-      this.action.some(a =>
-        new Matcher(applyContext(a, context)).match(action)
-      ) &&
-      this.resource.some(r =>
-        new Matcher(applyContext(r, context)).match(resource)
-      ) &&
       (conditionResolvers && this.condition && context
         ? Object.keys(this.condition).every(condition =>
             Object.keys(
               this.condition ? this.condition[condition] : {}
-            ).every(path =>
-              conditionResolvers[condition](
+            ).every(path =>{
+              if(this.condition){
+                const conditionValues=this.condition[condition][path]
+                if(conditionValues instanceof Array){
+                  return conditionValues.some(value=>conditionResolvers[condition](
+                    getValueFromPath(context, path),
+                    value
+                  ))
+                }
+                return conditionResolvers[condition](
+                  getValueFromPath(context, path),
+                  conditionValues
+                )
+              }
+              return conditionResolvers[condition](
                 getValueFromPath(context, path),
-                this.condition ? this.condition[condition][path] : ''
+                ''
               )
+            }
+              
             )
           )
         : true)
@@ -294,4 +233,5 @@ export class Statement {
   }
 }
 
-export { StatementConfig, ConditionResolver };
+
+export { Statement };
