+++ tests/ResourceBasedPolicy.ts
@@ -1,11 +1,11 @@
-import { IdentityBasedPolicy } from '../src/Role';
+import { ResourceBasedPolicy } from '../src/Policies';
 export default (): void => {
-  describe('IdentityBasedPolicy Class', () => {
+  describe('ResourceBasedPolicy Class', () => {
     describe('can create role', () => {
       it('don\'t throw an error', () => {
         expect(
           () =>
-            new IdentityBasedPolicy([
+            new ResourceBasedPolicy([
               {
                 resource: 'some:glob:*:string/wqweqw',
                 action: ['read', 'write'],
@@ -17,57 +17,61 @@ export default (): void => {
 
     describe('can match some actions and not others', () => {
       it('returns true or false', () => {
-        const role = new IdentityBasedPolicy([
+        const role = new ResourceBasedPolicy([
           {
+            principal: 'rogger',
             resource: ['books:horror:*'],
             action: ['read'],
           },
         ]);
-        expect(role.can('read', 'books:horror:The Call of Cthulhu')).toBe(true);
-        expect(role.can('write', 'books:horror:The Call of Cthulhu')).toBe(false);
+        expect(role.can('rogger','read', 'books:horror:The Call of Cthulhu')).toBe(true);
+        expect(role.can('andre','write', 'books:horror:The Call of Cthulhu')).toBe(false);
       });
     });
 
     describe('can match some resources and not others', () => {
       it('returns true or false', () => {
-        const role = new IdentityBasedPolicy([
+        const role = new ResourceBasedPolicy([
           {
+            principal: 'rogger',
             resource: ['books:horror:*'],
             action: ['read'],
           },
         ]);
-        expect(role.can('read', 'books:horror:The Call of Cthulhu')).toBe(true);
-        expect(role.can('read', 'books:fantasy:Brisingr')).toBe(false);
+        expect(role.can('rogger','read', 'books:horror:The Call of Cthulhu')).toBe(true);
+        expect(role.can('rogger','read', 'books:fantasy:Brisingr')).toBe(false);
       });
     });
 
     describe('can match based on context', () => {
       it('returns true or false', () => {
-        const role = new IdentityBasedPolicy([
+        const role = new ResourceBasedPolicy([
           {
+            principal:{id: 'rogger'},
             resource: ['secrets:${user.id}:*'],
             action: ['read', 'write'],
           },
           {
+            principal:{id: 'rogger'},
             resource: ['secrets:${user.bestfriends}:*'],
             action: 'read',
           },
         ]);
-        expect(role.can('read', 'secrets:123:sshhh', { user: { id: 123 } })).toBe(
+        expect(role.can('rogger','read', 'secrets:123:sshhh','id', { user: { id: 123 } })).toBe(
           true
         );
         expect(
-          role.can('write', 'secrets:123:sshhh', { user: { id: 123 } })
+          role.can('rogger','write', 'secrets:123:sshhh','id', { user: { id: 123 } })
         ).toBe(true);
-        expect(role.can('read', 'secrets:123:sshhh', { user: { id: 456 } })).toBe(
+        expect(role.can('rogger','read', 'secrets:123:sshhh','id', { user: { id: 456 } })).toBe(
           false
         );
         expect(
-          role.can('read', 'secrets:563:sshhh', {
+          role.can('rogger','read', 'secrets:563:sshhh','id', {
             user: { id: 456, bestfriends: [123, 563, 1211] },
           })
         ).toBe(true);
-        expect(role.can('write', 'secrets:123:sshhh')).toBe(false);
+        expect(role.can('rogger','write', 'secrets:123:sshhh')).toBe(false);
       });
     });
 
@@ -79,13 +83,15 @@ export default (): void => {
           },
         };
   
-        const role = new IdentityBasedPolicy(
+        const role = new ResourceBasedPolicy(
           [
             {
+              principal:{id: 'rogger'},
               resource: 'secrets:*',
               action: ['read', 'write'],
             },
             {
+              principal:{id: 'rogger'},
               resource: ['posts:${user.id}:*'],
               action: ['write', 'read', 'update'],
               condition: {
@@ -98,14 +104,14 @@ export default (): void => {
           conditions
         );
   
-        expect(role.can('read', 'secrets:123:sshhh', { user: { id: 123 } })).toBe(
+        expect(role.can('rogger','read', 'secrets:123:sshhh','id', { user: { id: 123 } })).toBe(
           true
         );
         expect(
-          role.can('write', 'posts:123:sshhh', { user: { id: 123, age: 17 } })
+          role.can('rogger','write', 'posts:123:sshhh', 'id',{ user: { id: 123, age: 17 } })
         ).toBe(false);
         expect(
-          role.can('read', 'posts:456:sshhh', { user: { id: 456, age: 19 } })
+          role.can('rogger','read', 'posts:456:sshhh', 'id',{ user: { id: 456, age: 19 } })
         ).toBe(true);
       });
     });
